requirements.txt
# Bibliotecas para web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Bibliotecas para processamento de PDF
PyPDF2==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0

# Bibliotecas para manipulação de dados
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Bibliotecas para servidor web
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Bibliotecas para banco de dados
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Utilitários
python-dotenv==1.0.0
tqdm==4.66.1 